refactor: revert unwanted changes
